add the functionality of adding/removing the UID from event list in entrant leave/accept page
