+++ code/app/src/main/java/com/example/myapplication/entrant/entrantSelectedPage.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.myapplication.entrant;
 
 import android.os.Bundle;
 
@@ -13,15 +13,12 @@ import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.TextView;
 
+import com.example.myapplication.R;
 import com.example.myapplication.model.Event;
 import com.example.myapplication.model.User;
 
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link entranteventdescription#newInstance} factory method to
- * create an instance of this fragment.
- */
-public class entranteventdescription extends Fragment {
+
+public class entrantSelectedPage extends Fragment {
 
     private Event event; // Store the event object
     private User user;
@@ -36,7 +33,7 @@ public class entranteventdescription extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.entranteventdescription, container, false);
+        View view = inflater.inflate(R.layout.entrantselectedpage, container, false);
 
         // Reference UI elements
         ImageButton backButton = view.findViewById(R.id.backArrowButton);
@@ -44,6 +41,8 @@ public class entranteventdescription extends Fragment {
         TextView organizerNameTextView = view.findViewById(R.id.organizerNameTextView);
         TextView eventDescriptionTextView = view.findViewById(R.id.eventDescriptionTextView);
         TextView eventDateTextView = view.findViewById(R.id.eventDateTextView);
+        Button acceptButton = view.findViewById(R.id.AcceptButton);
+        Button declineButton = view.findViewById(R.id.DeclineButton);
 
         // Populate UI with event details
         if (event != null) {
@@ -53,7 +52,7 @@ public class entranteventdescription extends Fragment {
         }
 
         // Back button to navigate to the event list
-        backButton.setOnClickListener(v -> Navigation.findNavController(v).navigate(R.id.action_entranteventdescription_to_entrantEventsList));
+        backButton.setOnClickListener(v -> Navigation.findNavController(v).navigate(R.id.action_entrantSelectedPage_to_entrantEventsList));
 
         // Expand description
         expandDescriptionButton.setOnClickListener(v -> {
@@ -66,6 +65,39 @@ public class entranteventdescription extends Fragment {
             }
         });
 
+        // Leave button to remove event from waitlist
+        acceptButton.setOnClickListener(v -> acceptEvent());
+
+        // Decline button functionality to remove event from selected list
+        declineButton.setOnClickListener(v -> declineEvent());
+
         return view;
+    }
+
+    // Accept event and add to the user's accepted events
+    private void acceptEvent() {
+        if (event != null) {
+            user.addAcceptedEvent(event.getEventId()); // Method to add to accepted events
+            user.removeSelectedEvent(event.getEventId()); // Remove from selected events list
+            event.removeFromSelectedlist(user.getDeviceID());
+            event.addToAcceptedlist(user.getDeviceID());
+            Log.d("EntrantSelectedPage", "Event accepted: " + event.getEventId());
+
+            // Navigate back to the event list after accepting
+            Navigation.findNavController(requireView()).navigate(R.id.action_entrantSelectedPage_to_entrantEventsList);
+        }
+    }
+
+    // Decline event and remove from selected events list
+    private void declineEvent() {
+        if (event != null) {
+            user.removeSelectedEvent(event.getEventId());
+            event.removeFromSelectedlist(user.getDeviceID());
+            event.addToDeclinedlist(user.getDeviceID());// Method to remove from selected events
+            Log.d("EntrantSelectedPage", "Event declined: " + event.getEventId());
+
+            // Navigate back to the event list after declining
+            Navigation.findNavController(requireView()).navigate(R.id.action_entrantSelectedPage_to_entrantEventsList);
+        }
     }
 }
