+++ code/app/src/main/java/com/example/myapplication/entrant/entranteventdescription.java
@@ -1,27 +1,26 @@
-package com.example.myapplication;
+package com.example.myapplication.entrant;
 
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
 import androidx.navigation.Navigation;
 
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.TextView;
 
+import com.example.myapplication.R;
 import com.example.myapplication.model.Event;
 import com.example.myapplication.model.User;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link entrantSelectedPage#newInstance} factory method to
+ * Use the {@link entranteventdescription#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class entrantSelectedPage extends Fragment {
+public class entranteventdescription extends Fragment {
 
     private Event event; // Store the event object
     private User user;
@@ -36,7 +35,7 @@ public class entrantSelectedPage extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.entrantselectedpage, container, false);
+        View view = inflater.inflate(R.layout.entranteventdescription, container, false);
 
         // Reference UI elements
         ImageButton backButton = view.findViewById(R.id.backArrowButton);
@@ -44,8 +43,6 @@ public class entrantSelectedPage extends Fragment {
         TextView organizerNameTextView = view.findViewById(R.id.organizerNameTextView);
         TextView eventDescriptionTextView = view.findViewById(R.id.eventDescriptionTextView);
         TextView eventDateTextView = view.findViewById(R.id.eventDateTextView);
-        Button acceptButton = view.findViewById(R.id.AcceptButton);
-        Button declineButton = view.findViewById(R.id.DeclineButton);
 
         // Populate UI with event details
         if (event != null) {
@@ -55,7 +52,7 @@ public class entrantSelectedPage extends Fragment {
         }
 
         // Back button to navigate to the event list
-        backButton.setOnClickListener(v -> Navigation.findNavController(v).navigate(R.id.action_entrantSelectedPage_to_entrantEventsList));
+        backButton.setOnClickListener(v -> Navigation.findNavController(v).navigate(R.id.action_entranteventdescription_to_entrantEventsList));
 
         // Expand description
         expandDescriptionButton.setOnClickListener(v -> {
@@ -68,35 +65,6 @@ public class entrantSelectedPage extends Fragment {
             }
         });
 
-        // Leave button to remove event from waitlist
-        acceptButton.setOnClickListener(v -> acceptEvent());
-
-        // Decline button functionality to remove event from selected list
-        declineButton.setOnClickListener(v -> declineEvent());
-
         return view;
-    }
-
-    // Accept event and add to the user's accepted events
-    private void acceptEvent() {
-        if (event != null) {
-            user.addAcceptedEvent(event.getEventId()); // Method to add to accepted events
-            user.removeSelectedEvent(event.getEventId()); // Remove from selected events list
-            Log.d("EntrantSelectedPage", "Event accepted: " + event.getEventId());
-
-            // Navigate back to the event list after accepting
-            Navigation.findNavController(requireView()).navigate(R.id.action_entrantSelectedPage_to_entrantEventsList);
-        }
-    }
-
-    // Decline event and remove from selected events list
-    private void declineEvent() {
-        if (event != null) {
-            user.removeSelectedEvent(event.getEventId()); // Method to remove from selected events
-            Log.d("EntrantSelectedPage", "Event declined: " + event.getEventId());
-
-            // Navigate back to the event list after declining
-            Navigation.findNavController(requireView()).navigate(R.id.action_entrantSelectedPage_to_entrantEventsList);
-        }
     }
 }
